% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{add_vec_to_mat_rows}
\alias{add_vec_to_mat_rows}
\title{Add a vector to every row of a matrix}
\usage{
add_vec_to_mat_rows(v, M)
}
\arguments{
\item{v}{numeric vector of length ncol(M).}

\item{M}{numeric matrix with length(v) columns.}
}
\value{
M with v added to each row.
}
\description{
Add a vector to every row of a matrix
}
\keyword{internal}
