% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analytic11F}
\alias{analytic11F}
\title{analytic11F}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Matrix with 5 rows that represent simulation runs and two columns that
         represent field response and a single experimental input.
}\if{html}{\out{</div>}}
}
\usage{
analytic11F
}
\description{
A simple analytic example with random simulated data, known functional
components, and specified error terms. It has two columns. The first
column represent the response, and second column represent a single
experimental input.
}
\keyword{datasets}
