% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analytic11S}
\alias{analytic11S}
\title{analytic11S}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Matrix with 15 rows that represent simulation runs and three columns that
         represent code response, experimental input, and calibration input.
}\if{html}{\out{</div>}}
}
\usage{
analytic11S
}
\description{
A simple analytic example with random simulated data, known functional
components, and specified error terms. It has two columns. The first
column represents the the simulated code response, second column
represents a single experimental input, and the third column represents a
single calibration input.
}
\keyword{datasets}
