% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ballField}
\alias{ballField}
\title{ballField}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Matrix or data frame with 63 rows that represent observations and two
         columns that represent the response variable (time) and experimental
         variable (height) in that order
}\if{html}{\out{</div>}}

\describe{
\item{time}{The time in seconds for the ball to hit the ground}
\item{height}{Height of the ball in meters before dropping}
}
}
\source{
\url{https://bobby.gramacy.com/surrogates/ball.csv}
}
\usage{
ballField
}
\description{
A wiffle ball is dropped from different heights (second column) and the
time it took to hit the ground is measured (first column).
}
\keyword{datasets}
