% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ballSim}
\alias{ballSim}
\title{ballSim}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Matrix or data frame with 100 rows that represent simulation runs and
         three columns that represent response variable (time), experimental
         input (height) and calibration input (gravity).
}\if{html}{\out{</div>}}

\describe{
\item{time}{The time in seconds for the ball to hit the ground}
\item{height}{height of the ball in meters before dropping}
\item{gravity}{Height of the ball (in meters) before dropping}
}
}
\usage{
ballSim
}
\description{
The time to hit the ground after dropping the wiffle ball is simulated
using a mathematical model and is implemented as a code. The data
represent code response (time) at different heights and gravity.
}
\keyword{datasets}
