% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prior.R
\name{build_prior}
\alias{build_prior}
\title{Building Prior Function}
\usage{
build_prior(dist, p1, p2)
}
\arguments{
\item{dist}{character string representing prior distribution family}

\item{p1}{double representing the first parameter of the prior distribution}

\item{p2}{double representing the second parameter of the prior distribution}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ function that given its input, x, computes the log of the chosen prior
         probability density function
}\if{html}{\out{</div>}}
}
\description{
\code{build_prior} uses user-given prior types and parameters to select and parametrize
the prior function. This function
setup prior functions to return log of prior distribution.
}
\examples{
# create a prior function for beta(2, 5). Note that the function compute log of priors and must be
# transformed
pr_fun <- build_prior(dist = "beta", p1 = 2, p2 = 5)$fun
round(exp(pr_fun(c(-1, 0, 0.1, 0.5, 0.9, 1, 2))), 3)

# create a prior function for a Uniform distribution with lower bound of -10, and upper bound of 10
pr_fun <- build_prior(dist = "uniform", p1 = -10, p2 = 10)$fun
round(exp(pr_fun(c(-11, -5, 0, 4, 10, 12))), 3)

# create a prior function for Gaussian distribution with mean of 1 and standard deviation of 2
pr_fun <- build_prior(dist = "normal", p1 = 1, p2 = 2)$fun
round(exp(pr_fun(c(-9, -5, -3, -1, 1, 3, 5, 7, 11))), 3)
}
