% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_plot.R
\name{calibration_plot}
\alias{calibration_plot}
\title{Draws a calibration plot}
\usage{
calibration_plot(y_true, y_prob, normalize = FALSE, smooths = F,
  conf = T)
}
\arguments{
\item{y_true}{Vector of know labels on the test set}

\item{y_prob}{Vector of probability predictions on the test set}
}
\description{
Returns a ggplot2 object of calibration plot
}
