% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol_cov.R
\name{chol_cov}
\alias{chol_cov}
\title{Compute the augmented covariance matrix}
\usage{
chol_cov(env)
}
\arguments{
\item{sim}{simulation data}

\item{field}{field data}

\item{phi}{model parameters/hyperprameters}
}
\value{
covariance matrix of augmented data matrix
}
\description{
Compute the augmented covariance matrix
}
