% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control}
\alias{control}
\title{Control the optional arguments of \code{calibrate()}}
\usage{
control(...)
}
\arguments{
\item{...}{Any of the optional arguments in \code{calibrate()} function}
}
\description{
Control the optional arguments of \code{calibrate()}
}
