% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation Between Rows of Two Matrices.}
\usage{
correlation(X, Y = NULL, theta, alpha)
}
\arguments{
\item{X}{matrix with at least two rows}

\item{Y}{matrix with same number of columns as X}

\item{theta}{vector of scale parameters (\eqn{\theta \in (0, \inf)}) with the same size as number of columns as X}

\item{alpha}{vector of smoothness parameters (\eqn{\alpha \in [1, 2]}) with the same size as number of columns as X}
}
\value{
a correlation matrix between the rows of given matrices
}
\description{
If only one matrix is provided, \code{correlation} would compute the correlation between
the rows of the given matrix. Each entry R\link{i, j} in the returned correlation matrix is
correlation between row i of X and row j of Y (X if Y is null) using given smoothness
and scale parameters. To compute correlation, separable power exponential correlation
structure is assumed and using Euclidean notion of distance between rows.
}
