% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{crps_sample}
\alias{crps_sample}
\title{empirical continuous ranked probability score for one ensemble vs one obs.
CRPS = mean|x - o| - 0.5 mean|x - x'|, the sample estimator over the ensemble.}
\usage{
crps_sample(ens, obs)
}
\description{
empirical continuous ranked probability score for one ensemble vs one obs.
CRPS = mean|x - o| - 0.5 mean|x - x'|, the sample estimator over the ensemble.
}
\keyword{internal}
