% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Df2}
\alias{Df2}
\title{Df2}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Matrix with 50 rows that represent simulation runs and three columns that
         represent field response and two experimental inputs.
}\if{html}{\out{</div>}}
}
\usage{
Df2
}
\description{
A simple analytic example with random simulated data, known functional
components, and specified error terms. It has three columns. The first
column represent the field response, and the next two columns represent
two experimental inputs.
}
\keyword{datasets}
