% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ds2}
\alias{Ds2}
\title{Ds2}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Matrix with 100 rows that represent simulation runs and five columns that
         represent code response, 2 experimental inputs, and 2 calibration inputs.
}\if{html}{\out{</div>}}
}
\usage{
Ds2
}
\description{
A simple analytic example with random simulated data, known functional
components, and specified error terms. It has five columns. The first
column represents the the simulated code response, the next two columns
(second and third) represent two experimental inputs, and the last two
columns (fourth and fifth) represent two calibration inputs.
}
\keyword{datasets}
