% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ensemble_long}
\alias{ensemble_long}
\title{long form of a prediction matrix joined to the observation meta.}
\usage{
ensemble_long(G, meta)
}
\description{
long form of a prediction matrix joined to the observation meta.
}
\keyword{internal}
