% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbc.R
\name{fbc}
\alias{fbc}
\title{\code{fbc} Class Constructor}
\usage{
fbc(x)
}
\arguments{
\item{x}{named list of length 7 containing "Phi", "estimates", "logPost", "priors",
"acceptance", "vars", "cache" as fields}
}
\value{
object of class \code{fbc} object with mentioned elements`
}
\description{
Creates an object of class \code{fbc} from a given list with correct elements
}
