% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoslem_test.R
\name{hoslem_test}
\alias{hoslem_test}
\title{Hosmer-Lemeshow Goodness of Fit (GOF) Test}
\usage{
hoslem_test(x, y, g = 10)
}
\arguments{
\item{x}{a numeric vector of observations, binary (0/1).}

\item{y}{expected values.}

\item{g}{number of bins to use to calculate quantile.}
}
\description{
Perform a Hosmer-Lemeshow Goodness of Fit (GOF) Test
}
\details{
The Hosmer-Lemeshow test is a statistical test for goodness of fit for logistic regression models.
}
