% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sipnet.R
\name{inject_traits}
\alias{inject_traits}
\title{overwrite the pft's calibrated trait columns with the proposal U.}
\usage{
inject_traits(baseline, soil_pft, U_traits)
}
\description{
overwrite the pft's calibrated trait columns with the proposal U.
}
\keyword{internal}
