% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{label_variable}
\alias{label_variable}
\title{readable variable label: the variable name with underscores as spaces.}
\usage{
label_variable(x)
}
\description{
readable variable label: the variable name with underscores as spaces.
}
\keyword{internal}
