% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Conditional (posterior) log likelihood of augmented response vector (z)}
\usage{
log_lik(covD)
}
\arguments{
\item{CovD}{augmented covariance matrix}
}
\value{
posterior log likelihood of augmented response given the augmented
           covariance matrix (its inverse and determinant) and residuals
}
\description{
Conditional (posterior) log likelihood of augmented response vector (z)
}
