% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{MH MCMC method for full Bayesian parameter estimation}
\usage{
mcmc(Nmcmc, nBurn, thining, phiInit, env)
}
\arguments{
\item{Nmcmc}{an integer representing total number of MCMC runs}

\item{nBurn}{an integer representing number of burn-in runs}

\item{thining}{an integer representing number of burn-in runs}

\item{phiInit}{a double vector of initial values for all parameters}

\item{env}{an environment of the calibrate function to be set as parent}
}
\value{
a KOH object that includes a matrix of all parameters' distribution
           and a vector of log likelihood updates of each usable MCMC runs
}
\description{
MH MCMC method for full Bayesian parameter estimation
}
