% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{param_label_unit}
\alias{param_label_unit}
\title{parameter label with its unit appended, for a facet strip or page label.}
\usage{
param_label_unit(x)
}
\description{
parameter label with its unit appended, for a facet strip or page label.
}
\keyword{internal}
