% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt_calibration.R
\name{platt_calibration}
\alias{platt_calibration}
\title{Platt scaling probability calibration}
\usage{
platt_calibration(p, y)
}
\arguments{
\item{p}{Estimated probabilities from fit model}

\item{y}{Binomial response variable used to fit model}
}
\value{
a vector of calibrated probabilities
}
\description{
Performs an platt scaling calibration of posterior probability to minimize log loss.
}
