% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fbc.R
\name{plot.fbc}
\alias{plot.fbc}
\title{Plot Calibration Model}
\usage{
\method{plot}{fbc}(
  x,
  parameter = "kappa",
  type = "density",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{\code{fbc} object (output of \code{calibrate()} function)}

\item{parameter}{character string representing the parameter class}

\item{type}{type of plotting}
}
\description{
Given a parameter, \code{plot.fbc()} can plot three types of plots:
\itemize{
\item "density": plots the prior and posterior density distribution of chosen parameter
\item "trace":   plots the trace of the chosen parameter over MCMC runs (afte thinning)
\item "fits":    plots the fit of field training data to calibrated model (using "MAP" method)
}
}
