% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmode.R
\name{pmode}
\alias{pmode}
\title{Compute Mode of a Vector}
\usage{
pmode(x, breaks = 20)
}
\arguments{
\item{x}{numeric vector}

\item{breaks}{integer representing number of bins to compute mode}
}
\value{
double represent estimated mode of
}
\description{
Compute Mode of a Vector
}
