% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fbc.R
\name{print.fbc}
\alias{print.fbc}
\title{Calibration Object Print}
\usage{
\method{print}{fbc}(x, ...)
}
\arguments{
\item{x}{\code{fbc} object}

\item{...}{other arguments}
}
\description{
\code{print.fbc()} prints the prior specification and parameter summary for all model parameters
}
\examples{
# build a simple calibration model and print the model estimates
cal1    <- calibrate(sim = analytic11S, field = analytic11F,
                     nMCMC = 5, nBurn = 0, thinning = 1,
                     kappaDist = "normalTr", kappaP1 = 0.5, kappaP2 = 0.25)
print(cal1)
}
