% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Calculate the calibration slope}
\usage{
slope(x, y)
}
\arguments{
\item{y}{Binomial response variable used to fit model}

\item{p}{Estimated probabilities from fit model}
}
\value{
the point estimate and 95\%ci of slope and intercept.
}
\description{
The calibration slope, which is calculated by regressing the observed outcome on the predicted probabilities.
}
