% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyField}
\alias{toyField}
\title{toyField}
\format{
A matrix with 63 rows and 2 column, 1st column (experimental
           variable: height) and 2nd column (response variable: time)
\describe{
  \item{height}{Height of the ball (in meters) before dropping}
  \item{time}{The time (in seconds) for the ball to fall to ground}
}
}
\source{
<https://bobby.gramacy.com/surrogates/ball.csv>
}
\usage{
toyField
}
\description{
Field observations of time of falling balls for different heights
}
\keyword{datasets}
