% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sipnet.R
\name{write_samples_rdata}
\alias{write_samples_rdata}
\title{write samples.Rdata in the object run.write.configs expects.}
\usage{
write_samples_rdata(ensemble.samples, file)
}
\description{
write samples.Rdata in the object run.write.configs expects.
}
\keyword{internal}
