% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibration of Computer Models}
\usage{
calibrate(
  sim,
  field,
  Nmcmc = 10000,
  nBurn = 1,
  thining = 1,
  kappa = "betashift2",
  k0 = 0.5,
  k1 = 1.2,
  k2 = 1.2,
  theta = "gamma",
  t0 = 0.5,
  t1 = 1.5,
  t2 = 0.1,
  alpha = "betashift",
  a0 = 1.75,
  a1 = 5,
  a2 = 2,
  sigma2 = "gamma",
  s0 = 1,
  s1 = 1.5,
  s2 = 1.5
)
}
\arguments{
\item{sim}{A \eqn{m \times (1+p+q)} matrix, representing simulation data,
where m, number of rows, is number of simulation runs, p is number of
experimental variables, and q is number of calibration variables. Plus one
represent the first column, which is the response variable.}

\item{field}{A \eqn{n \times (1+p)} matrix, representing field data, where n,
number of rows, is number of field observations and p is number of
experimental variables.}

\item{Nmcmc}{An integer for number of MCMC runs.}

\item{nBurn}{An integer for number of MCMC burn ins.}

\item{thining}{A double as MCMC thinning rate to remove auto-correlation.}

\item{kappa}{A string to specify the prior type of calibration parameters.}

\item{k0}{a scaler or vector of same length as number of calibration
parameters that represent initial value for calibration parameters
after scaling to (0, 1)}

\item{k1}{A double as first parameter of the chosen kappa distribution.}

\item{k2}{A double as second parameter of the chosen kappa distribution.}

\item{theta}{A string to specify the prior type for scale parameters.}

\item{t0}{a scaler in (0, Inf) representing initial values of theta (scale) parameters}

\item{t1}{A double as first parameter of the chosen theta distribution.}

\item{t2}{A double as second parameter of the chosen theta distribution.}

\item{alpha}{A string to specify the prior type for smoothness parameters.}

\item{a0}{a scaler in (1, 2) representing initial values of alpha (smoothness) parameters}

\item{a1}{A double as first parameter of the chosen alpha distribution.}

\item{a2}{A double as second parameter of the chosen alpha distribution.}

\item{sigma2}{A string to specify the prior type for variance parameters.}

\item{s0}{a scaler or vector of length three that represent initial value for marginal variance}

\item{s1}{A double as first parameter of the chosen sigma2 distribution.}

\item{s2}{A double as second parameter of the chosen sigma2 distribution.}
}
\value{
an output list containing:
\itemize{
\item estimates:      Point estimates for all parameters and hyperparameters
\item distributions:  \eqn{\frac{(Nmcmc - nBurn)}{thinning} \times k} parameters distribution matrix,
where k (number of columns) is total number of parameters
\item acceptance:     acceptance rate of proposed samples
\item logPost:        log posterior distribution of response
\item samples:        full unfiltered samples of MCMC runs
\item vars:           name of all parameters (based on below notation)
\item summary:        summary of all columns of distribution matrix
}
}
\description{
\code{calibrate} combines simulator and field data and uses given (or default) priors
and initial values of the model parameters, runs a full Bayesian Markov Chain Monte Carlo (MCMC)
algorithm and samples the posterior distribution of parameters.
}
\details{

}
\examples{
examples/calib
}
\references{
Kennedy MC, O’Hagan A (2001). “Bayesian calibration of computer models.”
\emph{Journal of the Royal Statistical Society}, \strong{Series B}, \strong{63(3)}, 425–464
\url{https://www2.stat.duke.edu/~fei/samsi/Oct_09/bayesian_calibration_of_computer_models.pdf}
}
