\name{calibration-package}
\alias{calibration-package}
\alias{calibration}
\docType{package}
\title{
\packageTitle{calibration}
}
\description{
\packageDescription{calibration}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{calibration}
\packageIndices{calibration}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{calibration}

Maintainer: \packageMaintainer{calibration}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
