% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbc_class.R
\docType{class}
\name{fbc-class}
\alias{fbc-class}
\title{fbc Class}
\description{
fbc Class
}
\section{Slots}{

\describe{
\item{\code{estimates}}{data.frame.}

\item{\code{Phi}}{data.frame.}

\item{\code{logPost}}{numeric.}

\item{\code{acceptance}}{numeric.}

\item{\code{vars}}{character.}

\item{\code{cache}}{environment.}
}}

