% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmode.R
\name{pmode}
\alias{pmode}
\title{Estimate Mode from a Sample}
\usage{
pmode(x, breaks = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{breaks}{integer representing number of bins to compute mode}
}
\value{
double represent estimated mode of given vector
}
\description{
\code{pmode} bins sample observations into equally spaced intervals and counts number of
observations in each bin. The bin with highest number of observations represents the
mode and mean of observations in that bin is returned as representative mode of the
vector.
}
\details{
Note that \code{pmode} estimates the mode value of the underlying distribution. It is
suitable to use when dealing with double vectors. Therefore, for integer vectors,
\code{pmode}, the estimated mode will differ from its exact value.
}
\examples{
# find the estimated mode of a vector
vec <- runif(100, 0, 10)
pmode(vec)
pmode(vec, breaks = 10)
}
