% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fbc.R
\name{predict.fbc}
\alias{predict.fbc}
\title{predict}
\usage{
predict.fbc(object, newdata)
}
\arguments{
\item{newdata}{a vector of new data}

\item{obj}{class fbc}
}
\value{
predictions
}
\description{
predict
}
