% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fbc.R
\name{print.fbc}
\alias{print.fbc}
\title{Calibration Object Print}
\usage{
\method{print}{fbc}(x)
}
\arguments{
\item{x}{\code{fbc} object}
}
\description{
\code{print.fbc()} prints the prior specification and parameter summary for all model parameters
}
