% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_builder.R
\name{prior_builder}
\alias{prior_builder}
\title{Building Prior Function}
\usage{
prior_builder(prior, p1, p2)
}
\arguments{
\item{prior}{character string representing prior distribution family}

\item{p1}{double representing the first parameter of the prior distribution}

\item{p2}{double representing the second parameter of the prior distribution}

\item{ln}{logical denoting whether the output prior function return log of value.
Default is True as it log priors are needed in computations and they
simplifies the computation.}
}
\value{
function that given its input, x, computes the log of the chosen prior
probability density function
}
\description{
\code{prior_builder} uses user-given prior types and parameters to select and parametrize
the prior function. This function
setup prior functions to return log of prior distribution.
}
