% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{read_cal_val_observations}
\alias{read_cal_val_observations}
\title{read the observations table from a cal-val-data checkout.}
\usage{
read_cal_val_observations(cal_val_dir)
}
\description{
read the observations table from a cal-val-data checkout.
}
\keyword{internal}
