% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyperPriors.R
\name{set_hyperPriors}
\alias{set_hyperPriors}
\title{Set Calibration Hyperparameter Priors}
\usage{
set_hyperPriors(
  thetaS = list(dist = "gamma", init = 0.5, p1 = 1.1, p2 = 0.1),
  alphaS = list(dist = "betashift", init = 1.7, p1 = 5, p2 = 2),
  thetaB = list(dist = "gamma", init = 0.5, p1 = 1.1, p2 = 0.1),
  alphaB = list(dist = "betashift", init = 1.7, p1 = 5, p2 = 2),
  sigma2S = list(dist = "inversegamma", init = 1, p1 = 0.01, p2 = 0.01),
  sigma2B = list(dist = "inversegamma", init = 1, p1 = 0.01, p2 = 0.01),
  sigma2E = list(dist = "inversegamma", init = 1, p1 = 0.01, p2 = 0.01),
  muB = list(dist = "uniform", init = 0, p1 = -5, p2 = 5)
)
}
\arguments{
\item{thetaS}{prior specification of simulator correlation scale}

\item{alphaS}{prior specification of simulator correlation smoothness}

\item{thetaB}{prior specification of bias-correction correlation scale}

\item{alphaB}{prior specification of bias-correction correlation smoothness}

\item{sigma2S}{prior specification of simulator marginal variance}

\item{sigma2B}{prior specification of bias-correction marginal variance}

\item{sigma2E}{prior specification of measurement error variance}

\item{muB}{prior specification of bias-correction constant mean}
}
\value{
nested list containing the prior specification for all calibration hyperparameters
}
\description{
\code{set_hyperPriors()} can be used to specify one or more hyperparameter priors. There are
8 classes of hyperparameters. Each class can be specified using a named list with four
fields:
\itemize{
\item \code{dist}: character string (vector of character strings) to specify the family of the
distributions. All common probability distribution are implemented.
\item \code{init}: initial value to start the MCMC algorithm
\item \code{p1}:   first distribution parameter
\item \code{p2}:   second distribution parameter
All hyperparameters have default values to work with variety of models. Therefore, it
is recommended to use this function without any argument (to use default values) to set
\code{hypers} argument of \code{calibrate()} function.
}
}
