% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fbc.R
\name{summary.fbc}
\alias{summary.fbc}
\title{Calibration Object Summary}
\usage{
\method{summary}{fbc}(object)
}
\arguments{
\item{object}{\code{fbc} object}
}
\value{
dataframe where rows represent model parameters and columns represent: mean,
median, mode, lwr50, upr50, lwr80, upr80, sd
}
\description{
\code{summary.fbc()} produces a dataframe of summary statistics for all model parameters
}
